requirements.txt
streamlit==1.32.0
openai==1.12.0
pymongo==4.6.1
python-dotenv==1.0.0
streamlit-webrtc==0.47.1
pydub==0.25.1
scipy==1.11.3
numpy==1.24.3
soundfile==0.12.1
av==10.0.0
sounddevice==0.4.6
aiortc==1.5.0
matplotlib==3.7.2
waitress==2.1.2
flask==2.2.3